# v2/openapi/fb-unified-openapi.yaml
openapi: 3.0.0
servers:
  - description: Local server
    url: http://0.0.0.0:8000
info:
  title: Fireblocks Provider Connectivity API v2
  version: 0.5.0
  x-logo:
    url: doc-assets/fireblocks-logo-lockup-network-navy.png
    altText: Fireblocks logo
  contact:
    name: Fireblocks
    url: https://fireblocks.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  description: 
    $ref: README.md
...
paths:
...
  /accounts/{accountId}/capabilities/transfers/deposits:
    get:
      operationId: getDepositMethods
      summary: Get list of supported deposit methods
      description: |
        Retrieves the list of supported deposit methods available for the specified account. Shows which deposit types, networks, and sources are supported for fund transfers.
      tags:
        - capabilities
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationStartingAfter'
        - $ref: '#/components/parameters/PaginationEndingBefore'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          description: List of deposit methods for account.
          content:
            application/json:
              schema:
                type: object
                required:
                  - capabilities
                properties:
                  capabilities:
                    type: array
                    items:
                      $ref: '#/components/schemas/DepositCapability'
              examples:
                allUsdcToBucket:
                  summary: USDCs deposit into a bucket or USD
                  description: |-
                    This response indicates that all USDC token on the various blockchains could be deposited either into the USDC bucket balance or into the USD balance. The USD balance can be specified directly while token and bucket balances must be defined in `GET /capabilities/assets` and referenced using their IDs.
                    The asset IDs in this example are the same IDs used in the examples of `GET /capabilities/assets`.
                  value:
                    capabilities:
                      - id: 29d7d25a-ed05-479a-9cde-d50b8203d3ec
                        deposit:
                          asset:
                            assetId: 360de0ad-9ba1-45d5-8074-22453f193d65
                          transferMethod: PublicBlockchain
                        balanceAsset:
                          nationalCurrencyCode: USD
                        addressCreationPolicy: CanCreate
                      - id: adfc2457-1bb7-40d0-9d46-3cfa2dd3a7e1
                        deposit:
                          asset:
                            assetId: 606bce6b-ff15-4704-9390-b9e32a6cfcff
                          transferMethod: PublicBlockchain
                        balanceAsset:
                          nationalCurrencyCode: USD
                        addressCreationPolicy: CanCreate
                      - id: eda3dbb8-ef94-400e-8a6b-2a38302b9505
                        deposit:
                          asset:
                            assetId: 4386cf4d-83b2-4410-96da-0d3919a45506
                          transferMethod: PublicBlockchain
                        balanceAsset:
                          nationalCurrencyCode: USD
                        addressCreationPolicy: CannotCreate
                      - id: b9f471f0-4920-464f-9c6a-97929cc31950
                        deposit:
                          asset:
                            assetId: 360de0ad-9ba1-45d5-8074-22453f193d65
                          transferMethod: PublicBlockchain
                        balanceAsset:
                          assetId: f0844d82-7097-4521-95bc-d843724a893e
                        addressCreationPolicy: CannotCreate
                      - id: 53a36de2-f36f-4301-8292-a0ecc1c32cda
                        deposit:
                          asset:
                            assetId: 606bce6b-ff15-4704-9390-b9e32a6cfcff
                          transferMethod: PublicBlockchain
                        balanceAsset:
                          assetId: f0844d82-7097-4521-95bc-d843724a893e
                        addressCreationPolicy: CanCreate
                      - id: dd59e6d1-c8ca-4499-8b3f-786ce417ed17
                        deposit:
                          asset:
                            assetId: 4386cf4d-83b2-4410-96da-0d3919a45506
                          transferMethod: PublicBlockchain
                        balanceAsset:
                          assetId: f0844d82-7097-4521-95bc-d843724a893e
                        addressCreationPolicy: CanCreate
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
...
  /accounts/{accountId}/transfers/deposits:
    get:
      operationId: getDeposits
      summary: Get list of deposits sorted by creation time in a descending order
      description: |
        Retrieves a paginated list of all deposit transactions for the specified account. Deposits are sorted by creation time in descending order and include all types of deposit operations.
      tags:
        - transfers
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationStartingAfter'
        - $ref: '#/components/parameters/PaginationEndingBefore'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          description: Deposits details.
          content:
            application/json:
              schema:
                type: object
                required:
                  - deposits
                properties:
                  deposits:
                    type: array
                    items:
                      $ref: '#/components/schemas/Deposit'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
  /accounts/{accountId}/transfers/deposits/{id}:
    get:
      operationId: getDepositDetails
      summary: Get deposit details
      description: |
        Retrieves detailed information about a specific deposit transaction, including status, amounts, source details, confirmation information, and processing details.
      tags:
        - transfers
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/EntityIdPathParam'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          description: List of deposits.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deposit'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
  /accounts/{accountId}/transfers/deposits/addresses:
    post:
      operationId: createDepositAddress
      summary: Create new deposit address
      description: |
        Creates a new deposit address for the specified account and asset. The generated address can be used to receive deposits for the specified cryptocurrency or token.
      tags:
        - transfersFiat
        - transfersBlockchain
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      requestBody:
        description: Deposit address details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositAddressCreationRequest'
      responses:
        '200':
          description: New deposit address created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositAddress'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
    get:
      operationId: getDepositAddresses
      summary: Get list of existing deposit addresses
      description: |
        Retrieves a paginated list of all deposit addresses associated with the specified account. Shows addresses for different cryptocurrencies and networks that can receive deposits.
      tags:
        - transfersFiat
        - transfersBlockchain
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationStartingAfter'
        - $ref: '#/components/parameters/PaginationEndingBefore'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          description: List of existing deposit addresses.
          content:
            application/json:
              schema:
                type: object
                required:
                  - addresses
                properties:
                  addresses:
                    type: array
                    items:
                      $ref: '#/components/schemas/DepositAddress'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
  /accounts/{accountId}/transfers/deposits/addresses/{id}:
    get:
      operationId: getDepositAddressDetails
      summary: Get details of a deposit address
      description: |
        Retrieves detailed information about a specific deposit address, including the address string, associated network, asset type, and usage metadata.
      tags:
        - transfersFiat
        - transfersBlockchain
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/EntityIdPathParam'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          description: New deposit address created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositAddress'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      operationId: disableDepositAddress
      summary: Disable a deposit address
      description: |
        Disables a specific deposit address, preventing it from receiving new deposits. Existing funds sent to the address may still be processed depending on timing and confirmation status.
      tags:
        - transfersFiat
        - transfersBlockchain
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/EntityIdPathParam'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          description: Deposit address disabled.
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  parameters:
...
    X-FBAPI-KEY:
      name: X-FBAPI-KEY
      in: header
      required: true
      description: API authentication key.
      schema:
        type: string
    X-FBAPI-TIMESTAMP:
      name: X-FBAPI-TIMESTAMP
      in: header
      required: true
      description: Request timestamp in milliseconds since Unix epoch.
      schema:
        type: number
    X-FBAPI-NONCE:
      name: X-FBAPI-NONCE
      in: header
      required: true
      description: Unique identifier of the request.
      schema:
        type: string
    X-FBAPI-SIGNATURE:
      name: X-FBAPI-SIGNATURE
      in: header
      required: true
      description: |-
        Request signature using the chosen cryptographic algorithm. The signature is to be calculated on concatenation of the following request fields in the specified order:
        - `X-FBAPI-TIMESTAMP` - `X-FBAPI-NONCE` - HTTP request method in upper case - Endpoint path, including the query parameters - Request body
      schema:
        type: string
    X-FBAPI-INITIATED-BY:
      name: X-FBAPI-INITIATED-BY
      in: header
      required: false
      description: Conditional. This header is provided only when the corresponding capability requirement is enabled.
      schema:
        type: string
...
    EntityIdPathParam:
      name: id
      in: path
      required: true
      description: Entity unique identifier.
      schema:
        type: string
    SubAccountIdPathParam:
      name: accountId
      in: path
      required: true
      description: Sub-account identifier.
      schema:
        type: string
        pattern: ^[A-Za-z0-9_-]+$
...
    PaginationLimit:
      name: limit
      in: query
      description: Maximum number of returned items.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 10
    PaginationStartingAfter:
      name: startingAfter
      in: query
      description: Object ID. Instructs to return the items immediately following this object and not including it. Cannot be used together with `endingBefore`.
      schema:
        type: string
    PaginationEndingBefore:
      name: endingBefore
      in: query
      description: Object ID. Instructs to return the items immediately preceding this object and not including it. Cannot be used together with `startingAfter`.
      schema:
        type: string
...
  responses:
    BadRequestResponse:
      description: Request could not be processed due to a client error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestError'
    UnauthorizedResponse:
      description: Request is unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
    ErrorResponse:
      description: Failed to process request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GeneralError'
...
  schemas:
    GeneralError:
      description: Should be returned only if there in no more specific object.
      type: object
      required:
        - message
        - errorType
      properties:
        message:
          description: Description of the error.
          type: string
        errorType:
          enum:
            - internal-error
            - not-found
    BadRequestErrorType:
      type: string
      enum:
        - schema-error
        - schema-property-error
        - bad-request
        - unknown-asset
        - unsupported-conversion
        - quote-not-ready
        - deposit-address-disabled
        - order-not-trading
        - idempotency-key-reuse
        - deposit-address-creation-not-allowed
        - unsupported-transfer-method
        - transfer-destination-not-allowed
        - unsupported-ramp-method
        - unsupported-source-asset
        - unsupported-destination-asset
        - amount-below-minimum
        - pii-missing
        - unsupported-external-source
        - unsupported-region
        - destination-not-whitelisted
    BadRequestError:
      description: |-
        Request could not be processed due to a client error.
        If the error is caused by a missing or wrong property in the request, `errorType` must have the value `schema-property-error` and the properties  `propertyName` and `requestPart` must be specified. For more general schema errors, the error type should be `schema-error` and `requestPart` must be specified. For all the other cases the error type should be `bad-request`.
      type: object
      required:
        - message
        - errorType
      properties:
        message:
          description: Description of the error.
          type: string
        errorType:
          $ref: '#/components/schemas/BadRequestErrorType'
        propertyName:
          description: |-
            Name of property that caused the error. By convention, should always start with a slash ("/"). If the property is nested, the path should be separated by slashes.
            This property is required if the error type is caused by a missing or wrong property in the request.
          type: string
          example: /topLevelProperyName/childPropertyName/thirdLevelPropertyName
        requestPart:
          $ref: '#/components/schemas/RequestPart'
    UnauthorizedError:
      description: Request is unauthorized
      type: object
      required:
        - message
        - errorType
      properties:
        message:
          description: Description of the error.
          type: string
        errorType:
          enum:
            - unauthorized
        propertyName:
          description: Name of property that caused the error.
          enum:
            - X-FBAPI-KEY
        requestPart:
          description: Request part where the error occurred.
          enum:
            - headers
    RequestPart:
      description: Request part where the error occurred.
      enum:
        - params
        - body
        - querystring
        - headers
...
    PositiveAmount:
      type: string
      pattern: ^\d+(\.\d+)?
...
    NationalCurrency:
      type: object
      required:
        - nationalCurrencyCode
      properties:
        nationalCurrencyCode:
          $ref: '#/components/schemas/NationalCurrencyCode'
        testAsset:
          type: boolean
    NativeCryptocurrency:
      type: object
      required:
        - cryptocurrencySymbol
      properties:
        blockchain:
          $ref: '#/components/schemas/Blockchain'
        cryptocurrencySymbol:
          $ref: '#/components/schemas/CryptocurrencySymbol'
        testAsset:
          type: boolean
    OtherAssetReference:
      type: object
      required:
        - assetId
      properties:
        assetId:
          description: ID of one of the assets returned in get-additional-assets.
          type: string
    CryptocurrencyReference:
      oneOf:
        - $ref: '#/components/schemas/NativeCryptocurrency'
        - $ref: '#/components/schemas/OtherAssetReference'
    PublicBlockchainCapability:
      type: object
      required:
        - asset
        - transferMethod
      properties:
        asset:
          $ref: '#/components/schemas/CryptocurrencyReference'
        transferMethod:
          enum:
            - PublicBlockchain
    PublicBlockchainAddress:
      allOf:
        - $ref: '#/components/schemas/PublicBlockchainCapability'
        - type: object
          required:
            - address
          properties:
            address:
              type: string
            addressTag:
              type: string
...
    AssetReference:
      oneOf:
        - $ref: '#/components/schemas/NationalCurrency'
        - $ref: '#/components/schemas/NativeCryptocurrency'
        - $ref: '#/components/schemas/OtherAssetReference'
...
    IbanCapability:
      type: object
      required:
        - asset
        - transferMethod
      properties:
        asset:
          $ref: '#/components/schemas/NationalCurrency'
        transferMethod:
          enum:
            - Iban
    IbanAddress:
      allOf:
        - $ref: '#/components/schemas/IbanCapability'
        - type: object
          required:
            - accountHolder
            - iban
          properties:
            accountHolder:
              $ref: '#/components/schemas/AccountHolderDetails'
            iban:
              $ref: '#/components/schemas/Iban'
...
    IbanTransfer:
      allOf:
        - $ref: '#/components/schemas/IbanTransferDestination'
        - type: object
          properties:
            referenceId:
              type: string
...
    OtherFiatTransfer:
      type: object
      description: Used for historic fiat transfers that do not fit a more specific transfer method.  It is impossible to create new deposits or withdrawal of this type using this API.
      required:
        - transferMethod
        - asset
        - amount
        - description
      properties:
        transferMethod:
          enum:
            - OtherFiat
        asset:
          $ref: '#/components/schemas/NationalCurrency'
        amount:
          $ref: '#/components/schemas/PositiveAmount'
        description:
          type: string
        tag:
          type: string
        referenceId:
          type: string
...
    PublicBlockchainTransaction:
      type: object
      allOf:
        - $ref: '#/components/schemas/PublicBlockchainTransactionDestination'
        - type: object
          properties:
            blockchainTxId:
              type: string
    PeerAccountTransferCapability:
      type: object
      required:
        - asset
        - transferMethod
      properties:
        asset:
          $ref: '#/components/schemas/AssetReference'
        transferMethod:
          enum:
            - PeerAccountTransfer
    InternalTransferMethod:
      type: object
      required:
        - asset
        - transferMethod
      properties:
        asset:
          $ref: '#/components/schemas/AssetReference'
        transferMethod:
          type: string
          enum:
            - InternalTransfer
    InternalTransferCapability:
      type: object
      allOf:
        - $ref: '#/components/schemas/InternalTransferMethod'
        - type: object
          required:
            - destinationPolicy
          properties:
            destinationPolicy:
              $ref: '#/components/schemas/InternalTransferDestinationPolicy'
    InternalTransferDestinationPolicy:
      enum:
        - AnyAccount
        - DirectParentAccount
      description: |-
        Indicates the transfer destination policy:
          `AnyAccount`: Transfer is possible to any account of the same user.
          `DirectParentAccount`: Transfer is possible for the direct parent only.
    PeerAccountTransferAddress:
      type: object
      allOf:
        - $ref: '#/components/schemas/PeerAccountTransferCapability'
        - type: object
          required:
            - accountId
          properties:
            accountId:
              type: string
    PeerAccountTransferDestination:
      type: object
      allOf:
        - $ref: '#/components/schemas/PeerAccountTransferAddress'
        - type: object
          required:
            - amount
          properties:
            amount:
              $ref: '#/components/schemas/PositiveAmount'
    InternalTransferAddress:
      type: object
      allOf:
        - $ref: '#/components/schemas/InternalTransferMethod'
        - type: object
          required:
            - accountId
          properties:
            accountId:
              type: string
...
    PeerAccountTransfer:
      type: object
      allOf:
        - $ref: '#/components/schemas/PeerAccountTransferDestination'
        - type: object
          properties:
            referenceId:
              type: string
    InternalTransfer:
      type: object
      allOf:
        - $ref: '#/components/schemas/InternalTransferDestination'
    TransferCapability:
      oneOf:
        - $ref: '#/components/schemas/PeerAccountTransferCapability'
        - $ref: '#/components/schemas/InternalTransferCapability'
        - $ref: '#/components/schemas/PublicBlockchainCapability'
        - $ref: '#/components/schemas/IbanCapability'
...
    Transfer:
      oneOf:
        - $ref: '#/components/schemas/PeerAccountTransfer'
        - $ref: '#/components/schemas/InternalTransfer'
        - $ref: '#/components/schemas/PublicBlockchainTransaction'
        - $ref: '#/components/schemas/IbanTransfer'
        - $ref: '#/components/schemas/OtherFiatTransfer'
    DepositDestination:
      oneOf:
        - $ref: '#/components/schemas/PublicBlockchainAddress'
        - $ref: '#/components/schemas/IbanAddress'
        - $ref: '#/components/schemas/PeerAccountTransferAddress'
        - $ref: '#/components/schemas/InternalTransferAddress'
...
    DepositAddressStatus:
      enum:
        - enabled
        - disabled
    DepositAddress:
      type: object
      required:
        - id
        - destination
        - status
      properties:
        id:
          type: string
        destination:
          $ref: '#/components/schemas/DepositDestination'
        status:
          $ref: '#/components/schemas/DepositAddressStatus'
    DepositAddressCreationRequest:
      type: object
      required:
        - idempotencyKey
        - transferMethod
      properties:
        idempotencyKey:
          type: string
        transferMethod:
          oneOf:
            - $ref: '#/components/schemas/PublicBlockchainCapability'
            - $ref: '#/components/schemas/IbanCapability'
    DepositCapability:
      type: object
      description: Capability to deposit to a balance asset using a specific transfer capability. `addressCreationPolicy` determines whether new deposit addresses could be requested for this deposit capability.
      required:
        - id
        - deposit
        - balanceAsset
        - addressCreationPolicy
      properties:
        id:
          type: string
        deposit:
          $ref: '#/components/schemas/TransferCapability'
        balanceAsset:
          $ref: '#/components/schemas/AssetReference'
        addressCreationPolicy:
          $ref: '#/components/schemas/DepositAddressCreationPolicy'
    Deposit:
      type: object
      required:
        - id
        - balanceAsset
        - balanceAmount
        - source
        - status
        - createdAt
      properties:
        id:
          type: string
        balanceAsset:
          $ref: '#/components/schemas/AssetReference'
        balanceAmount:
          $ref: '#/components/schemas/PositiveAmount'
        source:
          $ref: '#/components/schemas/Transfer'
        depositAddressId:
          type: string
          description: ID of the deposit address, created by posting /transfers/deposits/addresses
        status:
          $ref: '#/components/schemas/DepositStatus'
        createdAt:
          description: Time when the deposit was created.
          type: string
          format: date-time
        finalizedAt:
          description: Time when the deposit reached a final state.
          type: string
          format: date-time
...
    DepositStatus:
      enum:
        - pending
        - succeeded
        - failed
...
    DepositAddressCreationPolicy:
      enum:
        - CanCreate
        - CannotCreate
...
  securitySchemes: {}